Updates to makefiles to accomodate ast.C, etc.
In my recent work I added some static variables to the ast.h file.  As
they are static they need to be initialized in a *C file somewhere, so I
also added the file ast.C.   However, as I was moving quickly I only
added ast.C to one Makefile (the one called "Makefile") and didn't
bother with the others.  This commit fixes that issue.

I previously added ast.C to the relevant Makefile.am file in the
ExpressionSrc directory, so that is already covered.