Updates to FIND-AT, FIND-WHEN and WHEN Measures
This commit updates the handling of FIND-AT, FIND-WHEN and WHEN
measures so that all analysis modes (AC, DC, NOISE AND TRAN) use a
common code flow.  This also fixes TRAN_CONT for FIND-WHEN and WHEN
measure types, so that they works correctly.  For TRAN and TRAN_CONT
modes, a check that the interpolated WHEN time is within the FROM-TO
window was also added.  So, the control flow for all measure modes
is now:

  1) Has a WHEN condition occurred?  If so then generate the
interpolated WHEN time (or frequency or DC sweep value).  (Note:
this approach still supports the Xyce-specific "MINVAL hack".)

  2) If that interpolated WHEN time is within the FROM-TO window
then the RISE, FALL, CROSS count is updated.  If that updated count
meets the specified RISE/FALL/CROSS condition then the measure
is satisfied.

This new approach supports both continuous and non-continuous measures.
However, there may still be issues with:

  a) CROSS=0 and the measure value equaling the target value at the
first simulation step (e.g., at t=0).

  b) negative CROSS (or RISE or FALL) values for non-continuous
measures.

SON Bugs 1274 and 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304