Improve Support for Negative RISE/FALL/CROSS for WHEN
This commit improves support for negative RISE, FALL and CROSS values
(in particular for values <= -2) for FIND-WHEN and WHEN measures
for the non-continuous versions of those two measure types for AC,
DC, NOISE and TRAN measure modes.  That behavior is now the same as
for the "continuous" (e.g., AC_CONT) measures.  In particular:

  1) If val=-1 (or val=LAST) then only the measure value at the last
crossing will be returned.

  2) If val <= -2 then only the measure value at the (last-abs(val)+1)
crossing will be returned.  If there are fewer then abs(val) crossings
then the measure will "fail".

At this point, there may still be issues with CROSS=0 and the measure
value equaling the target value at the first simulation step (e.g.,
at t=0).

SON Bugs 1274 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313