Code Cleanup of FindWhen Measure Classes
This commit does some code cleanup on the FindWhen measure classes.
It moves some common code into an updateMeasureVars() function in
the FindWhenBase class.  It also renames a few local variables to make
it clearer when they are a time, frequency or DC sweep value. Finally,
it removes a now unused private variable (doneIfFound_).

There should be no change in functionality with this commit.

SON Bug 1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274