Support Continuous Mode and Negative RFC for DERIV
This commit makes several upgrades to the DERIV-AT and DERIV-WHEN
measure types for .AC, .DC, .NOISE and .TRAN analysis modes.  First,
the RISE, FALL and CROSS qualifiers are now supported for DERIV-WHEN
measures for AC, DC and NOISE.  Based on those changes, the
"continuous" measure modes (AC_CONT, DC_CONT, NOISE_CONT and TRAN_CONT)
are now also supported, as are RISE/FALL/CROSS values < -1, for all
measure modes.  So, the behaviors now are:

  1) If val >=1 (or is not specified) then the measure will return
the derivative value at the val-th crossing for non-continous modes.
It will return the derivative values at all crossings, starting
with <val> for continuous modes.

  2) If val=-1 (or val=LAST) then only the derivative value at the
last crossing will be returned.

  3) If val <= -2 then only the derivative value at the (last-abs(val)+1)
crossing will be returned.  If there are fewer then abs(val) crossings
then the measure will "fail".

  4) If the AT qualifier is specified then only one derivative value
will be returned.

Finally, this commit also fixes SON Bug 1304 for TRAN mode.  The
WHEN time is now correctly checked against the bounds for the
FROM-TO window, before the RISE/FALL/CROSS qualifier is updated
and checked.

SON Bugs 1274, 1304 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313