Clean-up Output Functions for FIND and WHEN Measures
This commit cleans up the printMeasureResult() functions in the
FindWhen and FindWhenCont classes.  The previous versions worked
(by accident) for successful FIND-AT measures.  These versions make
the criteria for successful FIND-AT, FIND-WHEN and WHEN measures
clearer.  The control flow is also now more similar to that in the
printVerboseMeasureResult() functions.

This commit also makes some other minor changes, mostly in
comment lines or in removing unneeded {} pairs.

SON Bugs 1274 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313