Updates to run new expression library in parallel
For the most part I had not been testing the new expression library in
parallel.  I finally started to do this last week.  There were a few
problems, although (so far) nothing that was hard to fix.

This commit fixes most of the problems that I know of.  The main issues
had to do with one of the "group" classes I created for use of the new
library with Xyce.

The new library pulls the values it needs out of Xyce via group classes.
Up until this commit, I had 2 main group classes:  the "main" expression
group, and the "outputs" expression group.  The "outputs" group was for
expressions on the .print line, and it mostly relies on the various Op
classes.

The "main" expression group was (in theory) to be used by everyone else.
But, it was not well suited to be used everywhere else, at least in
parallel.  I copied most of the functions in the main group from the
code that I wrote years ago to resolve .SENS objective functions.
This worked fine in serial, but in parallel this was a problem because
those functions used AllReduce operations.  For the .SENS objective
funcitons, this was fine (even in parallel) b/c the objective function
processing happens simultaneously on every processor.

But a lot of expressions which were using the main group are associated
with dependent device parameters.  Devices and their parameters are unique
to whatever processor they happen to be on, so AllReduce operations
will tend to hang.  So, we can't use that type of function in devices.

The quickest way to fix this was to create another group, called the
"deviceExpressionGroup".  In this group, it is assumed that everything
is local the processor, and it does not attempt any reduction
operations.  Also, any solution vector access is performed using LIDs
rather than GIDs.  As the LIDs are unique to each parameter of each
device, a unique device group is allocated  for every "dependent"
device parameter, and the LIDs specific to that parameter are
passed into that group after they are resolved by topology.

The other use of AllReduce in the main group was in the functions that
obtain global parameter values.  The main group was using the device mgr
function "getParamAndReduce".  But for the device group, this wasn't
necessary.

The new device group is derived from the main group, and just overloads
the functions that used AllReduce calls.  So, the device group overloads
the "double" and "std::complex<double>" versions of the getSolutionVal,
getCurrentVal and getGlobalParameterVal

There are still some efficiency issues I need to work on, but that will
happen in subsequent commits.