Fix for new expression table breakpoints w/.STEP
I discovered this when debugging one of the few remaining test failures,
which is for bug 1772, which ostensibly tests one of the older (most
deprecated, frankly) neutron models.

The test was failing, but for reasons I didn't expect.   This was a test
that specified table-based (time,value) sources for defects, and that
also used .STEP.

The table sources were specified inside of the reaction file, but inside the
source code they are set up as expressions.  When I ran this circuit
using the old expression library and the new expression library, they
matched perfectly on the first .STEP iteration, but diverged
significantly on the second .STEP iteration.  And, when I ran the
equivalent of the second iteration by itself (without a .STEP command),
the old and new expression libraries matched perfectly as well.

It turned out that the reason was breakpoint management.  For most
spice-style sources in the new expression library, breakpoints are
generated identically to how they are generated in the N_DEV_SourceData
classes.  Most of these sources don't need special management for .STEP.

The exception is the PWL source, which returns all of the breakpoints
the first time it is called, and then is "smart" enough to not do that
work on subsequent calls.   But, when moving from one .STEP iteration to
the next, this needs to be reset.  Otherwise all .STEP iterations other
than the first one get zero breakpoints.    I had set this up correctly
for the PWL source, but I had not done it for the tableOp operator in
the new expression library.  This commit fixes that mistake, and now
the bug 1772 test passes.

This required adding a lot of "setupBreakPoints" function calls
throughout the code.