Improvements to DC Mode DERIV Measures
This commit makes two changes to the DERIV measure for DC measure
mode.  First the initialized_ = true statement in the updateDC()
function is moved to the correct place.  This maintains the Xyce
convention that a FAILED measure, with an invalid set of FROM-TO
qualifiers, reports the start-end sweep values as the measure window
in stdout.

The  Base::isInvalidDCsweepWindow() function was also changed.  In
HSPICE, the FROM-TO window for a DC measure is only invalid if both
FROM and TO are given, and that FROM-TO range does not overlap with
the DC sweep range.  This change follows that convention, and was
needed so the direction of the FROM-TO window (when both are given)
does not matter relative to the direction of the DC sweep.  As an
example, these two measures should give the same answer:

  .MEASURE DC FROM1 DERIV V(3) WHEN V(3)=0.5 FROM=2.25 TO=2.75
  .MEASURE DC FROM2 DERIV V(3) WHEN V(3)=0.5 TO=2.25 FROM=2.75

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304