Improve fixes for Nan and Inf results.
This was prompted by the diode_cmc noise test, which was failing with
the new expression library.  It was failing b/c two of the outputs were
(essentially) infinity, and the old library was setting them to 1.0e+50
(keeping the sign the same) and the new library was instead setting them
to 1.0e+20, while forcing the sign to be positive.

Also, the old library wasn't being smart about real vs. complex numbers.

This commit fixes all of these issues, and now the diode_cmc_noise test
passes with new expression.