Small change to paramOp::output() function
This makes the paramOp::output() function more similar to
that of the funcOp::output() function, and now outputs a bit more
information.

The "output" functions are what are used by the "dumpExpressionTree"
function in the expression class.  They traverse the tree while
outputting infomation about each node along the way.

The paramOp and funcOp nodes are somewhat unique in that they are the
only node types that allow external trees to be attached.  For the
purposes of the funcOp::output and paramOp::output functions, it is
important to call the output() function at the top of the attached
tree.  The funcOp::output() function was doing this, and the
paramOp::output() function was not.  This was probably a symptom of
funcOp being much harder to debug, and me spending a lot more time with
it.  Anyway, paramOp::output() now calls paramNode_->output(), so this
issue is fixed in this commit.