Update Non-Free license/name for new copyright
When the new copyright for the "NonFree" version of Xyce was implemented, the
license changed (technically, we didn't have an official license for the
non-free variant, but it didn't matter since it was never distributed). Also,
the official name for this variant of Xyce is "XyceNF" (NF for non-free). The
term "non-free" is used in the GNU "libre" sense of free, and refers to the
variant of Xyce that has proprietary compact models.

Issue https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/51