Fix the outputs group to return correct FREQ
This issue was breaking two of the MEASURE_NOISE tests.  Easy enough to
fix.  The outputs group was getting frequency from the device manager,
and, as it turns out, the device manager didn't know anything about it
and unconditionally returned a zero.

So, now the outputs group gets FREQ from the ouptut manager instead, and
all is good.