Implement the getCurrentFreq function
This is a derived function, that was implemented in the AC class, but
not the NOISE class.  it is what gets called when one calls the
N_ANP_AnalysisManager::getCurrentFreq() function, ultimately.

The reason I am implementing it is related to my previous commit.  The
new expression library uses the "outputs" group for .print and .measure
related output expressions.  This output group was getting the FREQ
variable via the solverState/deviceManager, which in turn was getting it from the
analysis manager.   For AC calculations it had the correct value, but
for NOISE calculations it always returned a zero.  This issue was the
reason.

In my previous commit I changed the outputs group to get FREQ from the
output manager instead, so this is a somewhat moot commit.  But, I
wanted to make the code consistent.  Also, the "main" and "device"
expression groups still get FREQ via the device manager.