Improve Measure Warnings for Negative RFC Values
This commit extends the warning message measures for AC, NOISE and
TRAN mode measures that failed because an insufficient number of
rises, falls or crosses were found.  Those warning messages are now
correctly emitted for negative RISE, FALL and CROSS values.  So,
for example, if CROSS=LAST (which is the same as CROSS=-1) is given
on the .MEASURE line then a warning message is emitted to stdout if
no valid crosses were found.

SON Bugs 1274, 1313 and 1315
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1315