Make collapse errors actual fatal errors
Xyce/ADMS has always tested for errors in node collapse, which
mostly means an attempt to use incorrectly scoped variables or
bias-dependent variables in collapse conditionals.

Unfortunately, the error it emitted was not a fatal error, and so
"buildxyceplugin" would then try to keep going and compile the
resulting broken C++.

It should not do that.

This commit changes the "admst:error" that says we're aborting due to
collapse errors into an "admst:fatal" so it actually aborts.