Make newExpression::attachFunction/Param robust
As I was debugging the use of sdt and ddt inside of function trees
last week, I discovered that if a .func expression was un-parsable, but
then I attempted to attach the resulting expression to another
expression, there was no failsafe mechanism to prevent horrible things
from happening later.

In other words, if I had something like this:

.func F1(x) {sin((x)}  <-- missing the second paren

{10.0*f1(3.0)")} <-- fine by itelf, but calling a broken .func

when I tried to evaluate the second expression (the one invoking f1), I
would get a segfault.  The "attachFunction" didn't check if the pointer
from the .func was valid, which it would not be in this case.

I currently do NOT have syntax errors in expressions unconditionally
force the code to exit.  They just report these errors to the error
handler, which in Xyce means it will call Xyce::Report::UserError().

This commit adds several "if" statements to check if various RCP are
null or not.  If they are null, then don't attach.  Then, when the
expression is evaluated, it will throw an error saying that it isn't
properly resolved, because the parameterNode or functionNode pointers
are still null.