Test ensure newExpression::attachFunction robust
This commit adds at least one test, which is an error test.  If the
test runs w/out a seg fault, it passes.  It doesn't compare against a
gold standard value.

The specific case being tested is where an expression calls a .func, but
the expression on the RHS of the .func is not parsable.  So, it should
fail to be attached, and then should give a clean error message about
not being resolved when evaluated.