Applying the C++ "rule of 5" to astNode base class
This commit is an attempt to follow this guidance:
https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Rc-five

The rule of 5 is that since the semantics of copy, move and destruction
are closely related, if one of them is declared, the others need to be
as well.

For all but one of the ast node classes, I don't declare any of these,
and just let the compiler do its job.  But, for the base class, I
declared a virtual destructor.   This implies byt the rule of 5,
I have to do all the others.  But, I can explicitly declare
them all as "default", which is what I did:

In other words, this commit adds the following:

    virtual ~astNode() = default;
    astNode(const astNode&) = default;
    astNode& operator=(const astNode&) = default;
    astNode(astNode&&) = default;
    astNode& operator=(astNode&&) = default;