Begin work on shift/reduce conflicts
Prior to this commit Bison reported that there were 47 shift/reduce
conflicts in the lexer/parser.  I did a bit of reading this morning
about how to get rid of these and have started working on it.

As there are a bunch of them, I am doing this in stages.  This
particular commit doesn't eliminate very many of them.  It represents my
initial  work on the very first of these conflicts, which has to do with
rules starting with "TOK_WORD".  That particular token ("TOK_WORD") is
massively overused in the parser.  So, in this commit I got rid of many
uses of it.  Not enough to remove the conflict, but close.

More to come.