Fix most shift/reduce conflicts, except for POLY
This commit fixes a bunch of shift/reduce conflicts.  It intentionally
leaves the POLY expressions broken, because fixing that is a bit more
involved, and I already changed many things in this commit so I wanted
to save my work before heavily hacking the file more.

The issues resolved by this commit include:

(1) getting rid of 16 shift/reduce conflicts by noticing that one rule,
for the application of the TOK_NOT operator, was actually in the file
2x.  Once with TOK_NOT and once with '~', which means the same thing.
Getting rid of the '~' rule got rid of that problem.

(2) systematically working on the "nonempty_arglist" operator, which was
a source of many other shift/reduce conflicts.  The nonempty_arglist was
designed as a comma-separated list of operators of arbitrary length.
This is needed for things like function(arg1,arg2,arg3, ... ) and
various other operators such as spice_pulse.

The problem with nonempty_arglist is that as I originally wrote it, it
was impossible for the parser to tell the difference between a
single-element arglist and exp, which is the standard operator most
rules return and/or reduce.

I came up with a solution for it that involves adding more specific
context.   I now have an arglist object called "two_or_more_arglist",
which is not allowed to contain just one argument.  And I now have
another operator called "nonempty_func_arglist", which can either have a
single "exp" or a "two_or_more_arglist" as the input, but is required to
have a right paren as the next token.

Now, all (most?) operators that need an input of arbitrary
length comma-separated list of operators use the "nonempty_func_arglist"
object.  This has enough context in it that it doesn't cause conflicts.
The nice thing is that nearly all uses of comma-separate lists are
function arguments, so they are surrounded by parenthesis, so specifying
this context wasn't too hard.

I now have a similar problem to deal with for POLY, but it is a harder
problem so I am deferring it to a later commit.  POLY has two lists of
arbitrary length, with no delimeters.  The first list is just a no-comma
list of "exp" objects.  The second is a no-comma list of doubles
(coefficients).    Fortunately, I have a special token for the doubles
"TOK_POLYCOEF", so it should ultimately not be too difficult to handle
the list of "exp" that comes before it on the poly line.

But I haven't done that yet.  So, as of this commit, POLY is busted, and
all the POLY unit tests will fail.  The parser input file has a fair
amount of commented-out stuff in it.  Some of it is different
experiments I have already attempted on POLY.

Some of it is for a handful of TABLE specifications that I had implemented
but are apparently not tested via unit testing.  Possibly these
commented-out TABLE specifications will need to be restored once I run
the regression suite again.  And, once I do that, assuming these are
real specs that need to be supported, I need to add more TABLE unit test.

TABLE and POLY have been by far the hardest things to support, in terms
of parsing.  POLY, because it is somewhat context-free, and TABLE,
because there are so many variations (and also some of those variations
have tricky context).