Final(?) fix for shift/reduce errors
This commit fixes shift/reduce errors for POLY.

POLY was hard to parse without shift/reduce conflicts.  I took a
systematic approach, and broke the parsing into as many separate rules
as possible.  In particular, I wanted every rule associated with poly to
be unique to poly, and also to have at most 2 arguments, so the context
would be 100% known.    My approach was probably overkill, but it works
now, and so I'm doing a commit.

At this point, there are still a 2 or 3 table patterns commented out.
The ones that are commented out are not tested with my unit testing, so
either they are irrelevant, or I just forgot to create tests for those
patterns and will discover that I need to restore them after I run the
regression tests.

But, with that caveat, all the shift/reduce errors are gone, at least
for now.