Final, final shift/reduce conflict fix
In my last commit I got POLY working w/out shift/reduce conflicts.

At the end of that commit, I said that I'd left sevearl of the table
patterns/rules commented out.   As it turned out, there was only one
table rule commented out, not several.

Having that one rule commented out did not affect any unit tests
(indicating I need to create more table unit tests).  I was concerned
that leaving it out would break regression tests.  I have found one
newly-broken regression test (so far) but it doesn't appear to be
related to the missing table rule.

But, for completeness, this commit fixes the remaining table rule.  It
is no longer commented out, and I have revised the rule so that it no
longer produces any shift/reduce conflicts.

Also, as of this commit, the "nonempty_arglist" operator, which I had
previously used all over the place, is no longer being used by any rule.
It has been removed.  This operator was the cause of about 2/3 of the
shift/reduce errors.  Part of the problem was that it was a recursive
rule to produce a vector/list of arbitrary size, including
potentially size=1. So it was difficult for the parser to distinguish
a size=1 array from a single element.  They had the same pattern.  The
other part of the problem was that I was using nonempty_arglist in a
bunch of rules, which also had additional compounding shift/reduce
conflict problems.

Anyway, I think this is all fixed now, so one of the pre-pull-request
items is 99% done.  I just have to now double check what I broke in the
regression tests.