Comprehensive "changed" logic for new expression
In my previous commits, I implemented a preliminary version of the
"changed" logic that had been in place for the old expression library.
I had held off on implemting this concept for new expression, wanting to
get the new library 100% correct first, and to optimize after that.

As I noted in previous commits, the changed logic had two benefits for
device evaluation performance. (1) if an expression's inputs don't
change, then don't waste resources to re-evaluate it.  (2)  Most
expressions in the device package are attached to parameters.  When an
expression is re-evaluated, that is the first step.  The parameter which
depends upon the expression must be updated, and all the internal
variables of the device must be re-evaluated as well.  So, if an
expression inputs don't change, and thus the output of the expression is
unchanged, then the parameter doesn't need to be udpated, and all the
processParam function calls related to that parameter don't need to be
re-evaluated.  These things should only be executed when an expression
has legitimately changed.

In my previous commits I managed to implement a (roughly) 80% solution,
in which I mostly took care of (2) but not (1).  (2) was by far the most
costly, so it was a good place to start.

In this commit, I have now completely taken care of (1) as well, and
also fixed at least one mistake from my previous commit that caused a
new regresison test failure.  That failure is now gone.   I have also
refactored the "changed" logic in the device package so that it is much
easier to understand, and (at least under some circumstances) even more
efficient than it was with the old expresison library.  To some extent
this refactor was possible because of the superior design of the new
expression library.

With the old expression library, it was entirely the responsibility of
the calling code to determine if the inputs had changed.  As the new
library uses a "group" class to obtain values from Xyce, it assumes a
lot of this responsibility now, and determines for itself if the
inputs are any different than they were on the last call.    So,
with the most efficient options enabled, the new expresison library
will do a "pull" operation on the group, check them and then optionally
skip the evaluation, while telling the calling code that this happened.

But, this "pull" operation has some cost to it, albeit small. So it
behooves the calling code to avoid even that part if it can.   This
commit makes an attempt at that.

It does this by making use of the fact that the new expression
library can tell the calling code if it depends on specials
like "TIME", or if it has any dependence on solution variables
like V(1).   In the midst of a time-stepping loop, or a Newton
loop, if the expression depends on none of these things, then there
is no point in even checking the inputs for changes.  We know
already that there aren't any.

So, with my new changes, I can pre-filter the expression calls, and
save a little more calculation time.

At this point, I think the evaluation aspect of the new expression
library is generally as fast or faster than the old one.  It is still a
bit slower during the setup phase (i.e. parsing).  I plan to work
more on that next.