Add a debugging back-end for Xyce/ADMS
ADMS comes with a simple back-end called "vlatovla.xml" that takes a
Verilog-A input file, parses it, creates the ADMS data tree, then
dumps the data tree back out in Verilog-A with comments containing
additional data from the processing.  The output Verilog-A file should
be the same as the input, modulo formatting and these extra comments.

This original vlatovla.xml script is the basis for a new script I'm
starting to hack up, with more information output in a more readable
fashion.  I am getting ready to work on some Xyce/ADMS issues, and
they're all about fixing up incorrect data in the ADMS data tree.
This script should help me see what that data is and how my hackage on
the implicit rules is changing things.

To use this, one just does "admsXml -x -e $ADMSDIR/adms.implicit.xml
-e $ADMSDIR/xyceVLAToVLA.xml source.va" to process the source.va file.
It will output to a file named "vlatovla.va" (hard coded in this
version).  The output file is not formatted nicely --- no indentation
or anything.  For readability, one might want to run it through a good
verilog-a indenter (emacs does a good job in Verilog mode).

It is not strictly necessary to use Xyce's implicit rules file to use
this script, but there are some things in the Xyce implicit rules that
make analog functions with output variables more correct, specifically
dependencies of these output variables.  The consequence of this would
actually be visible in the output vlatovla.va file, which in this
version should show all variable interdependencies.  They'll be wrong
for analog function output variables without our implicit rules.

I will be expanding the information in the annotation comments a lot
over the coming development.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/54
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/52
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/48
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/40