Make analog function output vars have correct dependency
In commit 8611def, I had introduced a local adms.implicit.xml and
modified it to handle propagation of probe and variable dependencies
from analog function argument lists into the output variables (other
than the return value, which was already handled correctly) of analog
functions.

In doing so, I had neglected setting the "dependency" element of the
ADMS data tree on those variables, and so they were incorrectly being
set to "constant".  This led to the incorrect setting downstream of
the OPdependent element as well.

Xyce/ADMS does not currently use either of these bits of the data
tree, so this had no effect.  But in work I intend to do in the
future, these elements WILL be used, and so it is necessary for them
to be correct in analog function output variables.

This commit correctly sets "dependency", and therefore
OPdependent (which is set later by probing "dependency").

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/54