Minor performance enhancements. More to come.
This commit comes from an attempt to speed up the setup phase (i.e.
netlist parsing) when using the new expression library.  This
doesn't improve things very much, but is a step in the right direction.

During netlist parsing, the parser often has to query an expression to
determine what to do with it.  So, it asks if it has any dependencies
on voltage nodes, or global parameters, etc.  In the old expression
library, this query was often done via the "get_names" function,
which I always found a bit confusing.

In the new parser, I wrote a number of new functions to replace
get_names.  The new functions have a number of inefficiencies
in them, which I am starting to address.

In this commit, I mainly do two things.  (1) I combine a bunch of these
query functions into a single function.  (2) I fix up the "isConstant"
boolean function and use it in the N_UTL_Param class.  In both cases
there should be some minor performance improvment, as there will be
fewer allocations, push_backs and AST traversals.