Minor efficiency enhancement
In every expression I allocate a few things that aren't necessarily
needed.  This commit gets rid of the easiest one to get rid of.

In the constructor I always allocate a paramOp class, that I labeled as
"garbage".  The idea was that occasionally I would need a parameter
class that could serve as a single memory location for a variety of
temporary, useless objects.

In practice, I only use it in one place, processing function arguments.
And, of course, most expressions don't have function arguments.  So
unconditionally allocating it is silly.  So I've removed it from the
constructor and located it in the function argument code, so it is only
allocated if needed.

I did similar things with allocations for specials, such as TIME, FREQ,
etc.  I haven't gotten rid of those yet, but they are less likely to be
expensive allocations as the "specials" class is smaller and simpler.

In general, I think much of the difference between the setup costs of
new expression and old expression is that I'm doing more allocations in
new expression.  They are allocations of small things, and so individual
cases are no big deal, but when there are many thousands of them, they
add up.  So I am running profilers and looking for allocations that can
be done away with.  There are several more obvious ones that I can get
rid of I think.

With my recent optimizations, the LDRD circuit (45nm PDK, 2164 devices
mostly BSIM3 SOI, 9569 unknowns).  My setup time has improved a lot:

                        total setup  solvers
old library, 9/6/2020:    55s   40s      15s
new library, 9/1/2020:   121s   78s      43s
new library, 9/6/2020:    82s   62s      20s
new library, 9/7/2020:    76s   61s      15s
new library, 9/9/2020:    69s   53s      15s   <-- the latest

So for this circuit the new library has basically identical solver time
(thanks to my "changed" refactor) and is closer in setup time (53 vs
40).  This is a big improvement over a week ago.