Remove unnecessary Trilinos include from Xyce header
N_CIR_Xyce.h is installed by a "make install" of Xyce, and must be
included by any program that attempts to link to Xyce shared
libraries.

An unnecessary include of "Teuchos_RCP.hpp" in this header file meant
that anyone using Xyce shared libraries must ALSO have Trilinos
headers just to include this one file.

Since it is not used and causes problems for external linkage, I'm
removing it.